Rename: revlight to light
