+++ keyboards/lily58/light/info.json
