+++ keyboards/lily58/lily58.h
@@ -2,6 +2,6 @@
 
 #ifdef KEYBOARD_lily58_rev1
     #include "rev1.h"
-#elif KEYBOARD_lily58_revlight
-    #include "revlight.h"
+#elif KEYBOARD_lily58_light
+    #include "light.h"
 #endif
